+++ lib/rubocop/cop/rails/delegate.rb
@@ -29,17 +29,16 @@ module RuboCop
       #     foo.bar
       #   end
       class Delegate < Cop
-        include OnMethodDef
-
         MSG = 'Use `delegate` to define delegations.'
 
-        private
-
-        def on_method_def(node, method_name, args, body)
+        def on_def(node)
+          method_name, args, body = *node
           return unless trivial_delegate?(method_name, args, body)
           return if private_or_protected_delegation(node)
           add_offense(node, :keyword, MSG)
         end
+
+        private
 
         def autocorrect(node)
           method_name, args, body = *node